Refactored code for Country Section
